refactor(console): separate mdx usages from docs
