+++ packages/console/src/mdx-components/Step/index.module.scss
